remove follow feed & scripting because it is breaking across installs
I need to sort this out and make it replicable, but what I'll probably do in the end is
just a better Firehose feed off-site to firehose.ww0cj.radio or something and setup Echofeed there. If I run it on my own servers I'll probably get a better refresh rate anyways.